Factor out asGraphicsLayer utility function
